Added Eclipse plugin project
A custom “TransformationEngine” interface provides a method called
“transform” that accepts paths to source and target model files to be
transformed. A custom class that represents a transformation engine and
implements is called “MediniQvtTransformationEngine” and it implements
custom “TransformationEngine” interface. It relies on third-party
“EMFQvtProcessorImpl” class to evaluate QVT transformations. An abstract
custom class “ModelTransformationTemplate” conforms to a Template design
pattern. It encapsulates a skeleton of an algorithm that describes the
model transformation process. The algorithm consists of four steps:
• prepare source model placeholder
• prepare target model placeholder
• perform transformation
• obtain target model
These four steps are represented by methods called
“prepareSourceModelPlaceholder”, “prepareTargetModelPlaceholder”,
“performTransformation”, and “obtainTargetModelWithDiagram”. These
methods are invoked in a template method named as “transformSourceModel”
of a “ModelTransformationTemplate” class.
A custom concrete class that extends “ModelTransformationTemplate” is
called “MediniQvtRsaDomModelTransformation”. This class is a
customization of a class called “UsingMediniQVTfromApplications” which
is a class provided by medini QVT team to specify how to use the
transformation engine in a Java application. The “MediniQvt” part of a
class name means that the transformation process is performed by this
model transformation tool. The “Rsa” part of the name means that the
source model is provided by the Rational Software Architect and a
diagram in a target model is created using this tool also. The “Dom”
part in the name specifies that the DOM method of XML editing is used.
A custom class that encapsulates the command handling is called
“SampleHandler” which inherits from third-party “AbstractHandler” class.
It implements “execute” method that is called each time the plugin is
opened. The method makes use of a custom “TransformationDialog” that
inherits from third-party abstract “Dialog” class to visualize a window
with control elements for a plugin. Once an “OK” button of a dialog is
clicked a branch that checks the status of a dialog is executed inside
of “execute” method. This is the starting execution point to put the
transformation logic into. This is where the method
“transformSourceModel” of a “MediniQvtRsaDomModelTransformation” can be
called to trigger the transformation process.